+++ src/main/java/com/tablelog/tablelogback/domain/chat/controller/ChatController.java
@@ -1,12 +1,9 @@
-package com.tablelog.tablelogback.domain.chat;
+package com.tablelog.tablelogback.domain.chat.controller;
 
-import org.aspectj.bridge.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.event.EventListener;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.web.bind.annotation.RestController;
